Codigo corregido Laptop- falta android

+++ backend/src/main/java/colegio/comedor/interfaces/InterfazCriterio.java
@@ -9,6 +9,6 @@ import org.springframework.stereotype.Repository;
 import colegio.comedor.modelo.Criterio;
 
 @Repository
-public interface InterfazCriterio extends CrudRepository<Criterio,Long> {
+public interface InterfazCriterio extends CrudRepository<Criterio,Integer> {
 
 }
